# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/org_units:
    get:
      tags:
        - OrgUnits
      summary: List OrgUnits
      description: Get a paginated list of OrgUnits with optional filtering and sorting
      operationId: listOrgUnits
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgUnitList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - OrgUnits
      summary: Create OrgUnit
      description: Create a new OrgUnit
      operationId: createOrgUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrgUnitRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgUnit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/positions:
    get:
      tags:
        - Positions
      summary: List Positions
      description: Get a paginated list of Positions with optional filtering and sorting
      operationId: listPositions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Positions
      summary: Create Position
      description: Create a new Position
      operationId: createPosition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    OrgUnit:
      type: object
      required:
        - id
        - kind
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        code:
          nullable: true
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 200
        metadata:
          type: object
...
    CreateOrgUnitRequest:
      type: object
      required:
        - kind
        - name
        - metadata
      properties:
        kind:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        code:
          nullable: true
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 200
        metadata:
          type: object
...
    OrgUnitList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrgUnit'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Position:
      type: object
      required:
        - id
        - company_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        code:
          nullable: true
          type: string
          maxLength: 30
        description:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePositionRequest:
      type: object
      required:
        - company_id
        - name
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        code:
          nullable: true
          type: string
          maxLength: 30
        description:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PositionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []